+++ src/lsapy/functions/__init__.py
@@ -1,0 +1,4 @@
+"""Suitability Functions Module."""
+
+from lsapy.functions.discrete import *
+from lsapy.functions.membership import *