+++ CHANGELOG.rst
@@ -6,6 +6,19 @@ v0.1.0 (unreleased)
 -------------------
 Contributor to this version: Baptiste Hamon (@baptistehamon).
 
+Breaking changes
+^^^^^^^^^^^^^^^^
+* Changes in ``SuitabilityFunction`` (issue `#15 <https://github.com/baptistehamon/lsapy/issues/15>`_, PR `#23 <https://github.com/baptistehamon/lsapy/pull/23>`_):
+    * ``SuitabilityFunction`` is now an abstract class and must not be instantiated directly.
+    * ``func_method`` and ``func_params`` have been renamed to ``name`` and ``params`` respectively.
+    * ``SuitabilityFunction`` does not support the option to retrieve the implemented function from the ``func_method`` anymore. This is now implemented in the ``MembershipFunction`` class.
+    * ``map`` has been deprecated because of its redundancy with the ``__call__`` method. Changes will be permanent in LSAPy v0.1.0. Call the function directly instead.
+* ``MembershipSuitFunction`` has been renamed to ``MembershipFunction`` (issue `#15 <https://github.com/baptistehamon/lsapy/issues/15>`_, PR `#23 <https://github.com/baptistehamon/lsapy/pull/23>`_):
+    * ``func_method`` and ``func_params`` have been renamed to ``name`` and ``params`` respectively.
+    * ``fit`` method has been renamed to ``fit_functions``.
+* ``DiscreteSuitFunction`` has been renamed to ``DiscreteFunction`` (issue `#15 <https://github.com/baptistehamon/lsapy/issues/15>`_, PR `#23 <https://github.com/baptistehamon/lsapy/pull/23>`_):
+    * ``func_method`` and ``func_params`` have been renamed to ``name`` and ``params`` respectively.
+
 Internal changes
 ^^^^^^^^^^^^^^^^
 * Templates for requesting new features, asking question and submitting PR have been added (issue `#11 <https://github.com/baptistehamon/lsapy/issues/11>`_, PR `#12 <https://github.com/baptistehamon/lsapy/pull/12>`_).
@@ -14,6 +27,9 @@ Internal changes
 * `Pre-commit` has been setup and `ruff`, `codespell` and `numpydoc` hooks have been added (issue `#8 <https://github.com/baptistehamon/lsapy/issues/8>`_, PR `#18 <https://github.com/baptistehamon/lsapy/pull/18>`_/PR `#19 <https://github.com/baptistehamon/lsapy/pull/19>`_).
 * The autoupdate schedule of `pre-commit` has been set to weekly (PR `#21 <https://github.com/baptistehamon/lsapy/pull/21>`_)
 * The unused ``introduction.ipynb`` notebook has been removed (issue `#15 <https://github.com/baptistehamon/lsapy/issues/15>`_, PR `#20 <https://github.com/baptistehamon/lsapy/pull/20>`_).
+* The structure around ``SuitabilityFunction``, ``MembershipFunction``, and ``DiscreteFunction`` has been redesigned :
+    * The ``SuitabilityFunction`` has been moved to LSAPy ``core`` module.
+    * The ``MembershipFunction`` and ``DiscreteFunction`` have been moved to the ``function`` module and split into two different files: ``membership.py`` and ``discrete.py``.
 
 v0.1.0-dev2 (2025-05-25)
 ------------------------